# backend/pkg/api/payment_handlers.go
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"
	"time"

	"github.com/digi604/swarmmarket/backend/internal/common"
	"github.com/digi604/swarmmarket/backend/internal/payment"
	"github.com/digi604/swarmmarket/backend/internal/transaction"
	"github.com/digi604/swarmmarket/backend/internal/wallet"
	"github.com/digi604/swarmmarket/backend/pkg/middleware"
	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/stripe/stripe-go/v76/webhook"
)

// webhookTimestampTolerance is the maximum age of a webhook event we'll accept.
// This helps prevent replay attacks.
const webhookTimestampTolerance = 5 * time.Minute

// PaymentHandler handles payment HTTP requests.
type PaymentHandler struct {
	paymentService     *payment.Service
	transactionService *transaction.Service
	walletService      *wallet.Service
	webhookSecret      string
}

// NewPaymentHandler creates a new payment handler.
func NewPaymentHandler(paymentService *payment.Service, transactionService *transaction.Service, walletService *wallet.Service, webhookSecret string) *PaymentHandler {
	return &PaymentHandler{
		paymentService:     paymentService,
		transactionService: transactionService,
		walletService:      walletService,
		webhookSecret:      webhookSecret,
	}
}

// CreatePaymentRequest is the request body for creating a payment.
type CreatePaymentIntentRequest struct {
	TransactionID string `json:"transaction_id"`
	ReturnURL     string `json:"return_url"` // URL to redirect after payment confirmation
}

// CreatePaymentIntent handles POST /payments/intent - create payment intent for escrow.
func (h *PaymentHandler) CreatePaymentIntent(w http.ResponseWriter, r *http.Request) {
	if h.paymentService == nil {
		common.WriteError(w, http.StatusServiceUnavailable, common.ErrServiceUnavailable("payments are not configured"))
		return
	}

	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	var req CreatePaymentIntentRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	transactionID, err := uuid.Parse(req.TransactionID)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid transaction_id"))
		return
	}

	// Get transaction
	tx, err := h.transactionService.GetTransaction(r.Context(), transactionID)
	if err != nil {
		if err == transaction.ErrTransactionNotFound {
			common.WriteError(w, http.StatusNotFound, common.ErrNotFound("transaction not found"))
			return
		}
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get transaction"))
		return
	}

	// Verify buyer is making the payment
	if tx.BuyerID != agent.ID {
		common.WriteError(w, http.StatusForbidden, common.ErrForbidden("only the buyer can create a payment"))
		return
	}

	// Create payment intent
	result, err := h.paymentService.CreateEscrowPayment(r.Context(), &payment.CreatePaymentRequest{
		TransactionID: transactionID,
		BuyerID:       tx.BuyerID,
		SellerID:      tx.SellerID,
		Amount:        tx.Amount,
		Currency:      tx.Currency,
		ReturnURL:     req.ReturnURL,
	})
	if err != nil {
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to create payment"))
		return
	}

	common.WriteJSON(w, http.StatusCreated, result)
}

// GetPaymentStatus handles GET /payments/{paymentIntentId} - get payment status.
func (h *PaymentHandler) GetPaymentStatus(w http.ResponseWriter, r *http.Request) {
	if h.paymentService == nil {
		common.WriteError(w, http.StatusServiceUnavailable, common.ErrServiceUnavailable("payments are not configured"))
		return
	}

	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	paymentIntentID := chi.URLParam(r, "paymentIntentId")
	if paymentIntentID == "" {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("payment_intent_id is required"))
		return
	}

	status, err := h.paymentService.GetPaymentIntent(r.Context(), paymentIntentID)
	if err != nil {
		common.WriteError(w, http.StatusNotFound, common.ErrNotFound("payment not found"))
		return
	}

	common.WriteJSON(w, http.StatusOK, status)
}

// HandleWebhook handles POST /payments/webhook - Stripe webhook events.
func (h *PaymentHandler) HandleWebhook(w http.ResponseWriter, r *http.Request) {
	if h.paymentService == nil {
		common.WriteError(w, http.StatusServiceUnavailable, common.ErrServiceUnavailable("payments are not configured"))
		return
	}

	body, err := io.ReadAll(r.Body)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("failed to read body"))
		return
	}

	// Verify webhook signature with timestamp tolerance to prevent replay attacks
	sig := r.Header.Get("Stripe-Signature")
	event, err := webhook.ConstructEventWithOptions(body, sig, h.webhookSecret, webhook.ConstructEventOptions{
		IgnoreAPIVersionMismatch: true,
	})
	if err != nil {
		fmt.Printf("[Stripe Webhook] Signature verification error: %v\n", err)
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid signature"))
		return
	}

	// Additional timestamp check for replay attack prevention
	// Stripe's signature includes a timestamp, but we add an extra check
	eventTime := time.Unix(event.Created, 0)
	if time.Since(eventTime) > webhookTimestampTolerance {
		fmt.Printf("[Stripe Webhook] Event too old: %v\n", eventTime)
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("event expired"))
		return
	}

	// Handle different event types
	switch event.Type {
	case "payment_intent.succeeded":
		// Payment authorized - update escrow status
		h.handlePaymentSucceeded(r.Context(), event.Data.Raw)

	case "payment_intent.payment_failed":
		// Payment failed
		h.handlePaymentFailed(r.Context(), event.Data.Raw)

	case "charge.refunded":
		// Refund processed
		h.handleRefund(r.Context(), event.Data.Raw)
	}

	w.WriteHeader(http.StatusOK)
}

func (h *PaymentHandler) handlePaymentSucceeded(ctx context.Context, data []byte) {
	// Parse payment intent from webhook data
	var pi struct {
		ID       string            `json:"id"`
		Metadata map[string]string `json:"metadata"`
		Status   string            `json:"status"`
	}
	if err := json.Unmarshal(data, &pi); err != nil {
		return
	}

	// Check if this is a wallet deposit
	if pi.Metadata["type"] == "wallet_deposit" {
		if h.walletService != nil {
			if err := h.walletService.HandlePaymentIntentSucceeded(ctx, pi.ID); err != nil {
				fmt.Printf("[Stripe Webhook] Failed to handle wallet deposit: %v\n", err)
			}
		}
		return
	}

	// Handle transaction payment
	transactionIDStr, ok := pi.Metadata["transaction_id"]
	if !ok {
		return
	}

	transactionID, err := uuid.Parse(transactionIDStr)
	if err != nil {
		return
	}

	// For manual capture (escrow), "succeeded" means funds are captured
	// For automatic capture or when status is "requires_capture", mark as funded
	if pi.Status == "requires_capture" || pi.Status == "succeeded" {
		h.transactionService.ConfirmEscrowFunded(ctx, transactionID, pi.ID)
	}
}

func (h *PaymentHandler) handlePaymentFailed(ctx context.Context, data []byte) {
	var pi struct {
		ID               string            `json:"id"`
		Metadata         map[string]string `json:"metadata"`
		LastPaymentError *struct {
			Message string `json:"message"`
		} `json:"last_payment_error"`
	}
	if err := json.Unmarshal(data, &pi); err != nil {
		return
	}

	failureReason := "payment failed"
	if pi.LastPaymentError != nil && pi.LastPaymentError.Message != "" {
		failureReason = pi.LastPaymentError.Message
	}

	// Check if this is a wallet deposit
	if pi.Metadata["type"] == "wallet_deposit" {
		if h.walletService != nil {
			if err := h.walletService.HandlePaymentIntentFailed(ctx, pi.ID, failureReason); err != nil {
				fmt.Printf("[Stripe Webhook] Failed to handle wallet deposit failure: %v\n", err)
			}
		}
		return
	}

	// Handle transaction payment failure
	transactionIDStr, ok := pi.Metadata["transaction_id"]
	if !ok {
		return
	}

	transactionID, err := uuid.Parse(transactionIDStr)
	if err != nil {
		return
	}

	// Get transaction and publish failure event
	tx, err := h.transactionService.GetTransaction(ctx, transactionID)
	if err != nil {
		return
	}

	// Publish payment failed event for notification
	// The transaction stays in "pending" - buyer can retry
	h.transactionService.PublishPaymentFailed(ctx, tx, pi.ID, failureReason)
}

func (h *PaymentHandler) handleRefund(ctx context.Context, data []byte) {
	var charge struct {
		PaymentIntent string            `json:"payment_intent"`
		Metadata      map[string]string `json:"metadata"`
	}
	if err := json.Unmarshal(data, &charge); err != nil {
		return
	}

	transactionIDStr, ok := charge.Metadata["transaction_id"]
	if !ok {
		return
	}

	transactionID, err := uuid.Parse(transactionIDStr)
	if err != nil {
		return
	}

	// Mark transaction as refunded
	h.transactionService.RefundTransaction(ctx, transactionID)
}